create monsters table, some more work on monster spawning

+++ web/models/item_template.ex
@@ -12,7 +12,7 @@ defmodule ItemTemplate do
     field :hit_verbs,             {:array, :string}, default: []
     field :damage,                :string #json
     field :required_skills,       :string #json
-    field :speed,                 :float
+    field :speed,                 :decimal
     field :accuracy_skill,        :string
     field :ac,                    :integer
     field :uses,                  :integer
